refactor(archtest): Credential/Session/User 族迁 module-path-agnostic CellRule + 缩 baseline [#1633]
MOVE 9 个规则文件（18 条 INVARIANT）的检测逻辑从 _test.go → 非 test .go（Check* 函数），平台符号路径经 PlatformModulePath 派生，消除 bare github.com/ghbvf/gocell 字面量；_test.go 瘦身为 dogfood + 保留 meta/blind-spot 自检；module_path_funnel.baseline ratchet-down 9 行。行为零变更（verbatim MOVE + 恒等派生）。

规则均属 gocell-internal-layout（引用 cells/accesscore/internal/* 或 accesscore 路径 allowlist），外部 vacuous-pass，故迁为 importable Check* 但 NOT registered（与 errcode CheckErrorFirstAPI01 同形态）。

- USERREPO-CONFORMANCE-ENROLLMENT-01 B1 blind-spot（扫 "UserRepository" 字面量 reflect-bait）排除 tools/archtest/ 工具包——规则文件以该字符串标识目标类型，非生产 impl reflect bait；MOVE 前这些字面量在 _test.go（已排除）。
- gocognit 复杂度 ≤15 经行为保持的 helper 提取达成（无 //nolint）。

ref: golang.org/x/tools/go/analysis Analyzer
