+++ tools/archtest/cas_protocol_composition_root_test.go
@@ -51,25 +51,8 @@
 package archtest
 
 import (
-	"fmt"
-	"go/ast"
-	"strings"
 	"testing"
 )
-
-const (
-	casProtocolRuleID  = "CAS-PROTOCOL-COMPOSITION-ROOT-01"
-	casPkgImportPath   = "github.com/ghbvf/gocell/runtime/state/cas"
-	casOwnPkgRelPrefix = "runtime/state/cas/"
-	casOwnPkgRelExact  = "runtime/state/cas"
-)
-
-// casProtocolForbidden is the closed set of cas-package constructors banned
-// outside the composition root. (MustNewProtocol was deleted by B2-K-02;
-// only NewProtocol remains.)
-var casProtocolForbidden = map[string]struct{}{
-	"NewProtocol": {},
-}
 
 // TestCASProtocol_CompositionRootOnly enforces CAS-PROTOCOL-COMPOSITION-ROOT-01:
 // cas.NewProtocol may only be invoked from cmd/* (composition root) or
@@ -81,69 +64,6 @@ var casProtocolForbidden = map[string]struct{}{
 //   - examples/* each carry their own composition root; allowing them mirrors
 //     the AUTH-PLAN-04 / LAYER-09 carve-out for example projects.
 func TestCASProtocol_CompositionRootOnly(t *testing.T) {
-	diags := Run(t, Typed(
-		TypedOpts{Tests: false},
-		casProtocolProductionPatterns(),
-	),
-
-		scanCASProtocolViolations)
-
-	Report(t, casProtocolRuleID, diags)
-}
-
-// casProtocolProductionPatterns returns the package patterns scanned by the
-// production rule (cells / runtime / adapters).
-func casProtocolProductionPatterns() []string {
-	return []string{
-		"./cells/...",
-		"./runtime/...",
-		"./adapters/...",
-	}
-}
-
-// scanCASProtocolViolations walks every CallExpr in pass.Files, resolves the
-// callee to its (pkgPath, name) tuple via archtest.ResolvePackageRef, and flags
-// hits whose owning package is runtime/state/cas and whose name is in
-// casProtocolForbidden.
-//
-// Two file-level filters apply:
-//
-//   - _test.go suffix: defense-in-depth alongside TypedOpts{Tests: false}.
-//   - rel under "runtime/state/cas/": the package itself owns the constructor.
-//
-// Used by both TestCASProtocol_CompositionRootOnly (production scan, asserts
-// zero diagnostics) and TestCASProtocol_RedFixtureDetected (fixture scan,
-// asserts exactly 3 diagnostics across qualified / aliased / dot-import shapes).
-func scanCASProtocolViolations(p *Pass) []Diagnostic {
-	var out []Diagnostic
-	for _, file := range p.Files {
-		rel := p.Rel(file)
-		if strings.HasSuffix(rel, "_test.go") {
-			continue
-		}
-		// Exempt the cas package itself and its sub-packages.
-		if rel == casOwnPkgRelExact || strings.HasPrefix(rel, casOwnPkgRelPrefix) {
-			continue
-		}
-		EachInSubtree[ast.CallExpr](file, func(call *ast.CallExpr) {
-			pkgPath, name, ok := ResolvePackageRef(p.TypesInfo, call.Fun)
-			if !ok || pkgPath != casPkgImportPath {
-				return
-			}
-			if _, banned := casProtocolForbidden[name]; !banned {
-				return
-			}
-			line := p.Fset.Position(call.Pos()).Line
-			out = append(out, Diagnostic{
-				Rel:  rel,
-				Line: line,
-				Message: fmt.Sprintf(
-					"cas.%s must only be called from cmd/* (composition root) or runtime/state/cas/*; "+
-						"cells / runtime (non-cas) / adapters must consume an injected *cas.Protocol",
-					name,
-				),
-			})
-		})
-	}
-	return out
+	t.Parallel()
+	Report(t, ruleCASProtocolCompositionRoot01, CheckCASProtocolCompositionRoot01(t, ConfigForExternalCell{}))
 }
