+++ tools/archtest/user_repo_conformance_enrollment_test.go
@@ -36,10 +36,8 @@
 package archtest
 
 import (
-	"fmt"
 	"go/ast"
 	"go/types"
-	"sort"
 	"strings"
 	"testing"
 
@@ -47,17 +45,9 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/ghbvf/gocell/tools/internal/prodscan"
-	"github.com/ghbvf/gocell/tools/typesutil"
 )
 
 // INVARIANT: USERREPO-CONFORMANCE-ENROLLMENT-01
-
-const (
-	userRepoIfacePkg  = "github.com/ghbvf/gocell/cells/accesscore/internal/ports"
-	userRepoIfaceName = "UserRepository"
-	conformancePkg    = "github.com/ghbvf/gocell/cells/accesscore/internal/ports/conformance"
-	conformanceFunc   = "RunUserRepoConformance"
-)
 
 // TestUserRepoConformanceEnrollment enforces USERREPO-CONFORMANCE-ENROLLMENT-01:
 // every concrete type implementing ports.UserRepository in the production tree
@@ -68,117 +58,8 @@ func TestUserRepoConformanceEnrollment(t *testing.T) {
 	if testing.Short() {
 		t.Skip("skipping packages.Load-based archtest in -short mode")
 	}
-
-	root := findModuleRoot(t)
-
-	// ─── Step 1: resolve ports.UserRepository interface ─────────────────────
-	//
-	// The iface and the impl types MUST come from the same packages.Load
-	// invocation so that types.Implements uses pointer-identical *types.Named
-	// descriptors (cross-load comparisons are always false).
-	prodPatterns := prodscan.Patterns(root)
-	ifacePatterns := append([]string{"./cells/accesscore/internal/ports/..."}, prodPatterns...)
-
-	var userRepoIface *types.Interface
-	var implPkgs []*types.Package
-
-	_ = Run(t, Typed(TypedOpts{Tests: false, Tags: FlatNonDefaultTags()}, ifacePatterns),
-		func(p *Pass) []Diagnostic {
-			if p.Pkg == nil {
-				return nil
-			}
-
-			if p.Pkg.Path() == userRepoIfacePkg {
-				if obj := p.Pkg.Scope().Lookup(userRepoIfaceName); obj != nil {
-					if named, ok := obj.Type().(*types.Named); ok {
-						if iface, ok := named.Underlying().(*types.Interface); ok {
-							userRepoIface = iface.Complete()
-						}
-					}
-				}
-				return nil
-			}
-
-			implPkgs = append(implPkgs, p.Pkg)
-			return nil
-		})
-
-	require.NotNil(t, userRepoIface,
-		"USERREPO-CONFORMANCE-ENROLLMENT-01: failed to resolve ports.UserRepository interface; "+
-			"check import path %s", userRepoIfacePkg)
-
-	// ─── Step 2: collect all concrete implementations ───────────────────────
-	//
-	// implSet maps "pkg/path.TypeName" → true for every concrete exported type
-	// (struct or pointer-to-struct) that satisfies UserRepository.
-	// implPkgSet maps "pkg/path" → true for the owning package of each impl.
-	implSet := make(map[string]bool)
-	implPkgSet := make(map[string]bool) // pkg path → true
-
-	for _, pkg := range implPkgs {
-		if pkg == nil {
-			continue
-		}
-		collectUserRepoImpls(pkg, userRepoIface, implSet, implPkgSet)
-	}
-
-	// Regression guard: if zero impls found, the type-universe is broken.
-	require.NotEmpty(t, implSet,
-		"USERREPO-CONFORMANCE-ENROLLMENT-01: zero UserRepository implementations collected — "+
-			"likely a type-universe regression (iface and impls must share one packages.Load). "+
-			"Expect at least mem.UserRepository and postgres.PGUserRepo.")
-
-	// ─── Step 3: scan test corpus for RunUserRepoConformance call sites ─────
-	//
-	// A package is considered "enrolled" if any _test.go in it (or its test
-	// variant) contains a call to conformance.RunUserRepoConformance.
-	enrolledPkgs := make(map[string]bool) // pkg path → true
-
-	testPatterns := prodscan.Patterns(root)
-	_ = Run(t, Typed(TypedOpts{Tests: true, Tags: FlatNonDefaultTags()}, testPatterns),
-		func(p *Pass) []Diagnostic {
-			if p.Pkg == nil {
-				return nil
-			}
-			for _, f := range p.Files {
-				rel := p.Rel(f)
-				if !strings.HasSuffix(rel, "_test.go") {
-					continue
-				}
-				if hasConformanceCall(f, p.TypesInfo) {
-					pkgPath := canonicalPkgPath(p.Pkg.Path())
-					enrolledPkgs[pkgPath] = true
-				}
-			}
-			return nil
-		})
-
-	// ─── Step 4: flag unenrolled implementations ─────────────────────────────
-	var diags []Diagnostic
-	for implKey := range implSet {
-		// implKey is "pkg/path.TypeName"; extract pkg path.
-		dotIdx := strings.LastIndex(implKey, ".")
-		if dotIdx < 0 {
-			continue
-		}
-		pkgPath := implKey[:dotIdx]
-		if !enrolledPkgs[pkgPath] {
-			diags = append(diags, Diagnostic{
-				Rel:  implKey,
-				Line: 0,
-				Message: fmt.Sprintf(
-					"archtest: ports.UserRepository impl %q not enrolled in "+
-						"conformance.RunUserRepoConformance test call "+
-						"(USERREPO-CONFORMANCE-ENROLLMENT-01). "+
-						"Add a _test.go in package %s that calls "+
-						"conformance.RunUserRepoConformance(t, factory, features).",
-					implKey, pkgPath,
-				),
-			})
-		}
-	}
-	sort.Slice(diags, func(i, j int) bool { return diags[i].Rel < diags[j].Rel })
-	Report(t, "USERREPO-CONFORMANCE-ENROLLMENT-01", diags)
+	Report(t, ruleUserRepoConformanceEnrollment01,
+		CheckUserRepoConformanceEnrollment01(t, ConfigForExternalCell{BuildTags: FlatNonDefaultTags()}))
 }
 
 // TestUserRepoConformanceEnrollment_REDFixture verifies that the enrollment
@@ -290,7 +171,18 @@ func TestUserRepoConformanceEnrollment_ReverseBlindSpot_NoReflectImpl(t *testing
 		var out []Diagnostic
 		for _, f := range p.Files {
 			rel := p.Rel(f)
-			if strings.HasSuffix(rel, "_test.go") {
+			// Skip _test.go (a reflect-based impl would live in a production
+			// non-test file, not a test) and the archtest tooling package itself.
+			// tools/archtest rule files legitimately carry "UserRepository" as a
+			// string literal to identify the target type (userRepoIfaceName /
+			// userRepoType / the refreshGuardedMethods table) — that is the rule's
+			// own type-name spec, not a reflect-based implementation. Before #1633
+			// these literals lived in archtest _test.go (already excluded here); the
+			// rule-logic MOVE to non-test .go relocated them into production scope,
+			// so the exclusion is restated explicitly. A genuine reflect-based
+			// ports.UserRepository impl can only occur in cells/ runtime/ adapters/,
+			// all still scanned by ModuleScope below.
+			if strings.HasSuffix(rel, "_test.go") || strings.HasPrefix(rel, "tools/archtest/") {
 				continue
 			}
 			EachInSubtree[ast.BasicLit](f, func(lit *ast.BasicLit) {
@@ -314,49 +206,4 @@ func TestUserRepoConformanceEnrollment_ReverseBlindSpot_NoReflectImpl(t *testing
 
 	assert.Empty(t, diags,
 		"B1 reverse: no production non-test file should contain the string literal %q as reflect bait", userRepoIfaceName)
-}
-
-// ─── helpers ──────────────────────────────────────────────────────────────────
-
-// collectUserRepoImpls adds to implSet all exported concrete types in pkg that
-// implement UserRepository (directly or via pointer). Interface types are skipped.
-// implPkgSet receives the package path for each collected impl.
-func collectUserRepoImpls(pkg *types.Package, iface *types.Interface, implSet, implPkgSet map[string]bool) {
-	for _, name := range pkg.Scope().Names() {
-		obj, ok := pkg.Scope().Lookup(name).(*types.TypeName)
-		if !ok || !obj.Exported() {
-			continue
-		}
-		t := obj.Type()
-		// Skip interface types — only concrete types can be registered probers.
-		if _, isIface := t.Underlying().(*types.Interface); isIface {
-			continue
-		}
-		if typesutil.ImplementsInterface(t, iface) {
-			key := pkg.Path() + "." + name
-			implSet[key] = true
-			implPkgSet[pkg.Path()] = true
-		}
-	}
-}
-
-// hasConformanceCall returns true when file contains at least one call to
-// conformance.RunUserRepoConformance resolved via TypesInfo.
-func hasConformanceCall(file *ast.File, info *types.Info) bool {
-	if info == nil {
-		return false
-	}
-	_, ok := FindFirstInSubtree[ast.CallExpr](file, func(call *ast.CallExpr) bool {
-		pkgPath, name, resolved := ResolvePackageRef(info, call.Fun)
-		return resolved && pkgPath == conformancePkg && name == conformanceFunc
-	})
-	return ok
-}
-
-// canonicalPkgPath strips the "_test" or ".test" suffix from a test-variant
-// package path to obtain the canonical production package path.
-func canonicalPkgPath(path string) string {
-	path = strings.TrimSuffix(path, "_test")
-	path = strings.TrimSuffix(path, ".test")
-	return path
 }
